Package: geocodebr: Geolocalização De Endereços Brasileiros (Geocoding Brazilian Addresses)

Método simples e eficiente de geolocalizar dados no Brasil. O pacote é baseado em conjuntos de dados espaciais abertos de endereços brasileiros, utilizando como fonte principal o Cadastro Nacional de Endereços para Fins Estatísticos (CNEFE). O CNEFE é publicado pelo Instituto Brasileiro de Geografia e Estatística (IBGE), órgão oficial de estatísticas e geografia do Brasil. (A simple and efficient method for geolocating data in Brazil. The package is based on open spatial datasets of Brazilian addresses, primarily using the Cadastro Nacional de Endereços para Fins Estatísticos (CNEFE), published by the Instituto Brasileiro de Geografia e Estatística (IBGE), Brazil's official statistics and geography agency.)

See also

Author

Maintainer: Rafael H. M. Pereira rafa.pereira.br@gmail.com (ORCID)

Authors:

Daniel Herszenhut dhersz@gmail.com (ORCID)

Gabriel Garcia de Almeida (ORCID)

Other contributors:

Arthur Bazolli baz.arthur@gmail.com [contributor]

Pedro Milreu Cunha pedro.cunha@ipea.gov.br [contributor]

ITpS - Instituto Todos pela Saúde [funder]

Ipea - Instituto de Pesquisa Econômica Aplicada [copyright holder, funder]