Specify the columns describing the address fields

Creates a character vector specifying the columns that represent each address field in the addresses table.

Usage

setup_address_fields(
  logradouro = NULL,
  numero = NULL,
  cep = NULL,
  bairro = NULL,
  municipio = NULL,
  estado = NULL
)

Arguments

logradouro

A string. The name of the column representing the logradouro (street address) of the address. May be NULL if the field is not specified in the addresses table.

numero

A string. The name of the column representing the street number of the address. May be NULL if the field is not specified in the addresses table.

cep

A string. The name of the column representing the CEP (ZIP code) of the address. May be NULL if the field is not specified in the addresses table.

bairro

A string. The name of the column representing the neighborhood of the address. May be NULL if the field is not specified in the addresses table.

municipio

A string. The name of the column representing the city of the address. May be NULL if the field is not specified in the addresses table.

estado

A string. The name of the column representing the state of the address. May be NULL if the field is not specified in the addresses table.

Value

A character vector in which the names are the address fields and the values are the columns that represent them in the addresses table.

Examples

setup_address_fields(
  logradouro = "Nome_logradouro",
  numero = "Numero",
  cep = "CEP",
  bairro = "Bairro",
  municipio = "Cidade",
  estado = "UF"
)
#>        logradouro            numero               cep            bairro 
#> "Nome_logradouro"          "Numero"             "CEP"          "Bairro" 
#>         municipio            estado 
#>          "Cidade"              "UF"