Download the CNEFE data set

Downloads an enriched version of the CNEFE (National Registry of Addresses for Statistical Purposes, in portuguese) data set, purposefully built to be used with this package.

download_cnefe(state = "all", progress = TRUE, cache = TRUE)

Arguments

state

A character vector. The states whose CNEFE data should be downloaded. Either "all" (the default), in which case the data for all states is downloaded, or a vector with the states abbreviations (e.g. c("RJ", "DF") to download the data for Rio de Janeiro and the Federal District).

progress

A logical. Whether to display a download progress bar. Defaults to TRUE.

cache

Logical. Whether the function should read the data cached locally, which is much faster. Defaults to TRUE. The first time the user runs the function, censobr will download the file and store it locally so that the file only needs to be download once. If FALSE, the function will download the data again and overwrite the local file.

Value

A directory path where the data was saved.

Examples

# download CNEFE for a single state
download_cnefe(state = "AC", progress = FALSE)
#> [1] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=AC/part-0.parquet"

# download CNEFE for multiple states
download_cnefe(state = c("AC", "AL", "RJ"), progress = FALSE)
#> [1] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=AC/part-0.parquet"
#> [2] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=AL/part-0.parquet"
#> [3] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=RJ/part-0.parquet"

# download CNEFE for all states
download_cnefe(state = "all", progress = FALSE)
#>  [1] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=AC/part-0.parquet"
#>  [2] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=AL/part-0.parquet"
#>  [3] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=AM/part-0.parquet"
#>  [4] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=AP/part-0.parquet"
#>  [5] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=BA/part-0.parquet"
#>  [6] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=CE/part-0.parquet"
#>  [7] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=DF/part-0.parquet"
#>  [8] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=ES/part-0.parquet"
#>  [9] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=GO/part-0.parquet"
#> [10] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=MA/part-0.parquet"
#> [11] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=MG/part-0.parquet"
#> [12] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=MS/part-0.parquet"
#> [13] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=MT/part-0.parquet"
#> [14] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=PA/part-0.parquet"
#> [15] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=PB/part-0.parquet"
#> [16] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=PE/part-0.parquet"
#> [17] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=PI/part-0.parquet"
#> [18] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=PR/part-0.parquet"
#> [19] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=RJ/part-0.parquet"
#> [20] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=RN/part-0.parquet"
#> [21] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=RO/part-0.parquet"
#> [22] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=RR/part-0.parquet"
#> [23] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=RS/part-0.parquet"
#> [24] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=SC/part-0.parquet"
#> [25] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=SE/part-0.parquet"
#> [26] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=SP/part-0.parquet"
#> [27] "/home/runner/.cache/R/geocodebr/data_release_v0.1.0/estado=TO/part-0.parquet"