Download the geocodebr version of the CNEFE data set

Downloads an enriched version of the CNEFE data set that has been purposefully built to be used in the geocodebr package.

download_cnefe(abbrev_state = NULL, showProgress = TRUE, cache = TRUE)

Arguments

abbrev_state

description year

showProgress

Logical. Defaults to TRUE display download progress bar. The progress bar only reflects only the downloading time, not the time to load the data to memory.

cache

Logical. Whether the function should read the data cached locally, which is much faster. Defaults to TRUE. The first time the user runs the function, censobr will download the file and store it locally so that the file only needs to be download once. If FALSE, the function will download the data again and overwrite the local file.

Value

A directory path where the data was saved.

Examples

# download CNEFE for a single state
download_cnefe(abbrev_state = "AC",
               showProgress = FALSE)
#> [1] TRUE

# download CNEFE for multiple states
download_cnefe(abbrev_state = c("AC", "AL", "RJ"),
               showProgress = FALSE)
#> [1] TRUE

# download CNEFE for all states
download_cnefe(abbrev_state = "all",
               showProgress = FALSE)
#> Error in download_cnefe(abbrev_state = "all", showProgress = FALSE): Data currently only available for the states RO AC AM RR PA AP TO MA PI CE RN PB PE AL SE BA MG ES RJ SP PR SC RS MS MT GO DF