Download the CNEFE data set

Downloads an enriched version of the CNEFE (National Registry of Addresses for Statistical Purposes, in portuguese) data set, purposefully built to be used with this package.

Usage

download_cnefe(progress = TRUE, cache = TRUE)

Arguments

progress

A logical. Whether to display a download progress bar. Defaults to TRUE.

cache

A logical. Whether CNEFE data should be saved to/read from cache, reducing processing time in future calls. Defaults to TRUE. When FALSE, CNEFE data is downloaded to a temporary directory.

Value

Invisibly returns the path to the directory where the data was saved.

See also

Examples

download_cnefe(progress = FALSE)