Geocode Brazilian addresses

Geocodes Brazilian addresses based on CNEFE data. Addresses must be passed as a data frame in which each column describes one address field (street name, street number, neighborhood, etc). The input addresses are matched with CNEFE following 12 different case patterns. For more info, please see the Details section.

Usage

geocode(
  addresses_table,
  address_fields = setup_address_fields(),
  n_cores = 1,
  progress = TRUE,
  cache = TRUE
)

Arguments

addresses_table

A data frame. The addresses to be geocoded. Each column must represent an address field.

address_fields

A character vector. The correspondence between each address field and the name of the column that describes it in addresses_table. The setup_address_fields() function helps creating this vector and performs some assertions on the input. Address fields passed as NULL are ignored and the function must receive at least one non-null field. If manually creating the vector, please note that the vector names should be the same names used in the setup_address_fields() parameters.

n_cores

A number. The number of cores to be used in parallel execution. Defaults to 1.

progress

A logical. Whether to display progress bars when downloading CNEFE data and when geocoding the addresses. Defaults to TRUE.

cache

A logical. Whether CNEFE data should be saved to/read from cache, reducing processing time in future calls. Defaults to TRUE. When FALSE, CNEFE data is downloaded to a temporary directory.

Value

Returns the data frame passed in addresses_table with the latitude (lat) and longitude (lon) of each matched address, as well as another column (match_type) indicating the match level with which the address was matched.

Details

The input addresses are deterministically matched with CNEFE following 12 different case patterns. The case used to match each address is indicated in the output by the match_type column. Despite the case, the function always calculates the average latitude and longitude of all addresses in CNEFE that match the input address. For example, for the strictest case, in which the function finds a perfect match for the address, think of a building with several apartments that match the same street address: their coordinates will differ very slightly, and we take their average. In the other hand, for the loosest case, in which only the state and the city match, we take the city-wide average coordinates, which tend to favor more densely populated areas. The columns considered in each of the 12 different match types are described below:

Case 01: estado, município, logradouro, número, cep e bairro;

Case 02: estado, município, logradouro, número e cep;

Case 03: estado, município, logradouro, número e bairro;

Case 04: estado, município, logradouro, cep e bairro;

Case 05: estado, município, logradouro e número;

Case 06: estado, município, logradouro e cep;

Case 07: estado, município, logradouro e bairro;

Case 08: estado, município e logradouro;

Case 09: estado, município, cep e bairro;

Case 10: estado, município e cep;

Case 11: estado, município e bairro;

Case 12: estado, município.

Examples

data_path <- system.file("extdata/sample_1.csv", package = "geocodebr")
input_df <- read.csv(data_path)

fields <- setup_address_fields(
  logradouro = "nm_logradouro",
  numero = "Numero",
  cep = "Cep",
  bairro = "Bairro",
  municipio = "nm_municipio",
  estado = "nm_uf"
)

df <- geocode(input_df, address_fields = fields)
#> Matched addresses: 0/28 ■                                  0% - Looking for cas…
#> Matched addresses: 13/28 ■■■■■■■■■■■■■■■                   46% - Looking for ca…
#> Matched addresses: 16/28 ■■■■■■■■■■■■■■■■■■                57% - Looking for ca…
#> Matched addresses: 17/28 ■■■■■■■■■■■■■■■■■■■               61% - Looking for ca…
#> Matched addresses: 20/28 ■■■■■■■■■■■■■■■■■■■■■■            71% - Looking for ca…
#> Matched addresses: 24/28 ■■■■■■■■■■■■■■■■■■■■■■■■■■■       86% - Looking for ca…
#> Matched addresses: 24/28 ■■■■■■■■■■■■■■■■■■■■■■■■■■■       86% - Looking for ca…
#> Matched addresses: 24/28 ■■■■■■■■■■■■■■■■■■■■■■■■■■■       86% - Looking for ca…
#> Matched addresses: 24/28 ■■■■■■■■■■■■■■■■■■■■■■■■■■■       86% - Looking for ca…
#> Matched addresses: 25/28 ■■■■■■■■■■■■■■■■■■■■■■■■■■■■      89% - Looking for ca…
#> Matched addresses: 26/28 ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     93% - Looking for ca…
#> Matched addresses: 27/28 ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    96% - Looking for ca…
#> Matched addresses: 28/28 ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% - Done! 
#>