Download microdata of population records from Brazil's census

Download microdata of population records from Brazil's census. Data collected in the sample component of the questionnaire.

geocode(
  input_table,
  logradouro = NULL,
  numero = NULL,
  complemento = NULL,
  cep = NULL,
  bairro = NULL,
  municipio = NULL,
  estado = NULL,
  showProgress = TRUE,
  output_simple = TRUE,
  ncores = NULL,
  cache = TRUE
)

Arguments

showProgress

Logical. Defaults to TRUE display download progress bar. The progress bar only reflects only the downloading time, not the time to load the data to memory.

output_simple

Logic. Defaults to TRUE

ncores

Number of cores to be used in parallel execution. Defaults to the number of available cores minus 1.

cache

Logical. Whether the function should read the data cached locally, which is much faster. Defaults to TRUE. The first time the user runs the function, censobr will download the file and store it locally so that the file only needs to be download once. If FALSE, the function will download the data again and overwrite the local file.

year

Numeric. Year of reference in the format yyyy. Defaults to 2010.

columns

String. A vector of column names to keep. The rest of the columns are not read. Defaults to NULL and read all columns.

add_labels

Character. Whether the function should add labels to the responses of categorical variables. When add_labels = "pt", the function adds labels in Portuguese. Defaults to NULL.

Value

An arrow Dataset or a "data.frame" object.

Examples

library(geocodebr)

# open input data
data_path <- system.file("extdata/sample_1.csv", package = "geocodebr")
input_df <- read.csv(data_path)

df <- geocodebr::geocode_duck(
  input_table = input_df,
  logradouro = "nm_logradouro",
  numero = "Numero",
  complemento = "Complemento",
  cep = "Cep",
  bairro = "Bairro",
  municipio = "nm_municipio",
  estado = "nm_uf"
  )
#> Error: 'geocode_duck' is not an exported object from 'namespace:geocodebr'