geocodebr: Geocode Brazilian Addresses with CNEFE

Simple and efficient way to geocode data in Brazil. The package provides functions to geocode from addresses to spatial coordinates and to reverse-geocode from spatial coordinates to addresses. The package is built on top of open spatial data sets of Brazilian addresses, mainly using the National Registry of Addresses for Statistical Purposes (English for *Cadastro Nacional de Endereços para Fins Estatísticos*, CNEFE). CNEFE is published by the Brazilian official statistics and geography office, IBGE (https://www.ibge.gov.br/estatisticas/sociais/populacao/38734-cadastro-nacional-de-enderecos-para-fins-estatisticos.html).

See also

Author

Maintainer: Daniel Herszenhut dhersz@gmail.com (ORCID)

Authors:

Rafael H. M. Pereira (ORCID)

Other contributors:

Ipea - Institute for Applied Economic Research [copyright holder, funder]